requisito 1 - refatorando usando context
