+++ src/context/Context.jsx
@@ -1,0 +1,5 @@
+import { createContext } from 'react';
+
+const AppContext = createContext();
+
+export default AppContext;
