+++ src/pages/Main.jsx
@@ -1,15 +1,11 @@
-import React, { useEffect, useState } from 'react';
+import React from 'react';
+import Provider from '../context/Provider';
 import Table from '../components/Table';
 
 export default function Main() {
-  const [planets, setPlanets] = useState([]);
-  useEffect(() => {
-    const fetchAPI = async () => {
-      const request = await fetch('https://swapi-trybe.herokuapp.com/api/planets/');
-      const { results } = await request.json();
-      setPlanets(results);
-    };
-    fetchAPI();
-  }, []);
-  return (<Table planets={ planets } />);
+  return (
+    <Provider>
+      <Table />
+    </Provider>
+  );
 }
